requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
bert-score==0.3.12
bitarray==2.8.3
bitsandbytes==0.41.1
blis==0.7.9
blosc2==2.0.0
boto3==1.28.26
botocore==1.31.26
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.8
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cmake==3.26.4
colorama==0.4.6
confection==0.1.0
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1686613057838/work
cycler==0.11.0
cymem==2.0.7
Cython==3.0.5
decord==0.6.0
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.2
fairscale==0.4.13
fastapi==0.99.1
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fschat==0.2.18
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==0.4.6
gradio==3.35.2
gradio_client==0.2.7
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.15.1
Hydra==2.5
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
inexactsearch==1.0.2
iopath==0.1.10
Jinja2==3.1.2
jiwer==3.0.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
lemminflect==0.2.1
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==16.0.6
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdown2==2.4.9
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
munch==4.0.0
murmurhash==1.0.9
networkx==3.1
nh3==0.2.14
nltk==3.8.1
numexpr==2.8.6
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
orjson==3.9.1
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pathy==0.10.2
peft==0.3.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
portalocker==2.7.0
preshed==3.0.8
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-transformers==1.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.7
referencing==0.29.1
regex==2023.6.3
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
rich==13.4.2
rouge==1.0.1
rouge-score==0.0.4
rpds-py==0.8.8
rsa==4.9
s3transfer==0.6.1
sacrebleu==2.3.1
sacremoses==0.0.43
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.27.1
setproctitle==1.3.2
shortuuid==1.0.11
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soundex==1.1.3
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stanfordcorenlp==3.9.1.1
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
timm==0.9.2
tokenizers==0.10.3
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1
torchcam==0.3.2
torchdata==0.7.1
torchmetrics==1.2.0
torchvision==0.13.1+cu113
tqdm==4.65.0
transformers==4.12.5
triton==2.1.0
typer==0.9.0
typing_extensions @ file:///croot/typing_extensions_1686601443660/work
tzdata==2023.3
uc-micro-py==1.0.2
urllib3 @ file:///croot/urllib3_1686163155763/work
uvicorn==0.22.0
wandb==0.15.5
wasabi==1.1.2
wavedrom==2.0.3.post3
wcwidth==0.2.6
webdataset==0.2.48
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
